Wait for process to stop
Fixes rspec --order rand:19986

+++ spec/spec_system_helper.rb
@@ -131,10 +131,15 @@ shared_context "system" do
 
     def wait
       Process.wait(@pid)
+    rescue Errno::ESRCH
     end
 
     def stop
+      return unless @pid
       Process.kill("TERM", @pid)
+      wait
+    ensure
+      @pid = nil
     end
 
     def bin
